+++ docs/source/changelog.rst
@@ -5,6 +5,7 @@ v2.0.0 (2022-4-22)
 * Added surrogate model explanation method
 * Added support for attributing ECFP, MACCS fingerprints, rdkit descriptors and plotting them
 * Example notebooks for new method
+* Added more randomization approaches to STONED to improve mixing
 
 
 v1.1.0 (2022-5-2)